adds domain flow router object
